refactor: remove deprecated test helper files
- Deleted internal/testutils/deprecated folder
- Confirmed all functionality moved to card_helpers.go
- Fixed code review item #3 that identified unnecessary file
