+++ .gitignore
@@ -56,8 +56,4 @@ CONSULT-REQUEST.md
 CONSULTANT-PLAN.md
 *-PLAN.md
 **/glance.md
-
-# Development philosophy and documentation
-DEVELOPMENT_PHILOSOPHY.md
-DEVELOPMENT_PHILOSOPHY_APPENDIX_GO.md
 docs/prompts
