+++ DEVELOPMENT_PHILOSOPHY.md
@@ -1,0 +1,1 @@
+/Users/phaedrus/Development/codex/docs/DEVELOPMENT_PHILOSOPHY.md
